Was able to move all scripts into one file
