+++ test/util/fallbackSandBox.js
@@ -5,9 +5,10 @@ function fallbackDependancySandBox (appDir) {
   const path = require('path')
   const testSrc = path.resolve(__dirname, '../../test')
   const { execSync } = require('child_process')
-  const repoList = ['repo1', 'repo2', 'repo3', 'repo4', 'repo5', 'repo6', 'repo7', 'repo8',
-    'repo9', 'repo10', 'repo11', 'repo12', 'repo16', 'repo17', 'repo18', 'repo19', 'repo20',
-    'repo21', 'repo22', 'repo23', 'repo24', 'repo25', 'repo26', 'repo27'
+  const repoList = [
+    'repo1', 'repo2', 'repo3', 'repo4', 'repo5', 'repo6', 'repo7', 'repo8', 'repo9',
+    'repo10', 'repo11', 'repo12', 'repo13', 'repo14', 'repo15', 'repo16', 'repo17', 'repo18',
+    'repo19', 'repo20', 'repo21', 'repo22', 'repo23', 'repo24', 'repo25', 'repo26', 'repo27'
   ]
 
   if (!appDir) {
@@ -380,6 +381,127 @@ function fallbackDependancySandBox (appDir) {
       requires: true,
       packages: {}
     }
+    const repo13Package = {
+      devDependencies: {
+        'fallback-dependencies': '../../../'
+      },
+      fallbackDependencies: {
+        dir: 'lib',
+        reposFile: 'reposFile.json'
+      },
+      scripts: {
+        postinstall: 'node node_modules/fallback-dependencies/fallback-dependencies.js'
+      }
+    }
+    const repo13PackageLocked = {
+      name: 'repo13',
+      lockfileVersion: 3,
+      requires: true,
+      packages: {
+        '': {
+          hasInstallScript: true,
+          devDependencies: {
+            'fallback-dependencies': '../../../'
+          }
+        },
+        '../../..': {
+          version: '0.1.0',
+          dev: true,
+          license: 'CC-BY-13.0',
+          devDependencies: {}
+        },
+        'node_modules/fallback-dependencies': {
+          resolved: '../../..',
+          link: true
+        }
+      }
+    }
+    const repo13FileData = {
+      'fallback-deps-test-repo-30': [
+        '../../../repos/repo30'
+      ]
+    }
+    const repo14Package = {
+      devDependencies: {
+        'fallback-dependencies': '../../../../../../'
+      },
+      fallbackDependencies: {
+        dir: 'lib',
+        repos: {
+          'fallback-deps-test-repo-15': '../../../../../repos/repo15'
+        }
+      },
+      scripts: {
+        postinstall: 'node node_modules/fallback-dependencies/fallback-dependencies.js'
+      }
+    }
+    const repo14PackageLocked = {
+      name: 'repo13',
+      lockfileVersion: 3,
+      requires: true,
+      packages: {
+        '': {
+          hasInstallScript: true,
+          devDependencies: {
+            'fallback-dependencies': '../../../../../'
+          }
+        },
+        '../../../../..': {
+          version: '0.1.0',
+          dev: true,
+          license: 'CC-BY-13.0',
+          devDependencies: {}
+        },
+        'node_modules/fallback-dependencies': {
+          resolved: '../../../../..',
+          link: true
+        }
+      }
+    }
+    const repo15Package = {}
+    const repo15PackageLocked = {
+      name: 'repo15',
+      lockfileVersion: 3,
+      requires: true,
+      packages: {}
+    }
+    const repo30Package = {
+      devDependencies: {
+        'fallback-dependencies': '../../../../../../'
+      },
+      fallbackDependencies: {
+        dir: 'lib',
+        repos: {
+          'fallback-deps-test-repo-15': '../../../../../repos/repo15'
+        }
+      },
+      scripts: {
+        postinstall: 'node node_modules/fallback-dependencies/fallback-dependencies.js'
+      }
+    }
+    const repo30PackageLocked = {
+      name: 'repo13',
+      lockfileVersion: 3,
+      requires: true,
+      packages: {
+        '': {
+          hasInstallScript: true,
+          devDependencies: {
+            'fallback-dependencies': '../../../../../'
+          }
+        },
+        '../../../../..': {
+          version: '0.1.0',
+          dev: true,
+          license: 'CC-BY-13.0',
+          devDependencies: {}
+        },
+        'node_modules/fallback-dependencies': {
+          resolved: '../../../../..',
+          link: true
+        }
+      }
+    }
     const repo16Package = {
       devDependencies: {
         'fallback-dependencies': '../../../'
@@ -720,6 +842,7 @@ function fallbackDependancySandBox (appDir) {
     const packageList = [
       [repo1Package, repo1PackageLocked], [repo2Package, repo2PackageLocked], [repo3Package, repo3PackageLocked], [repo4Package, repo4PackageLocked], [repo5Package, repo5PackageLocked], [repo6Package, repo6PackageLocked],
       [repo7Package, repo7PackageLocked], [repo8Package, repo8PackageLocked], [repo9Package, repo9PackageLocked], [repo10Package, repo10PackageLocked], [repo11Package, repo11PackageLocked], [repo12Package, repo12PackageLocked],
+      [repo13Package, repo13PackageLocked], [repo14Package, repo14PackageLocked], [repo15Package, repo15PackageLocked],
       [repo16Package, repo16PackageLocked], [repo17Package, repo17PackageLocked], [repo18Package, repo18PackageLocked], [repo19Package, repo19PackageLocked], [repo20Package, repo20PackageLocked], [repo21Package, repo21PackageLocked],
       [repo22Package, repo22PackageLocked], [repo23Package, repo23PackageLocked], [repo24Package, repo24PackageLocked], [repo25Package, repo25PackageLocked], [repo26Package, repo26PackageLocked], [repo27Package, repo27PackageLocked]
     ]
@@ -756,6 +879,9 @@ function fallbackDependancySandBox (appDir) {
       if (repoList[id] === 'repo10') {
         fs.writeFileSync(`${testSrc}/clones/repo10/reposFile.json`, JSON.stringify(repo10FileData))
       }
+      if (repoList[id] === 'repo13') {
+        fs.writeFileSync(`${testSrc}/clones/repo13/reposFile.json`, JSON.stringify(repo13FileData))
+      }
       if (repoList[id] === 'repo16') {
         fs.writeFileSync(`${testSrc}/clones/repo16/reposFile.json`, JSON.stringify(repo16FileData))
       }
@@ -787,6 +913,9 @@ function fallbackDependancySandBox (appDir) {
         stdio: 'pipe', // hide output from git
         cwd: path.resolve(`${testSrc}/clones/${repoList[id]}`, '') // where we're cloning the repo to
       })
+      if (repoList[id] === 'repo13') {
+        fs.rmSync(path.resolve(`${testSrc}/clones/repo14` + '/.git/config'), { recursive: true, force: true })
+      }
     }
 
     execSync('npm ci', {
@@ -812,6 +941,32 @@ function fallbackDependancySandBox (appDir) {
     execSync('npm ci', {
       stdio: 'pipe', // hide output from git
       cwd: path.resolve(`${testSrc}/clones/repo10`, '') // where we're cloning the repo to
+    })
+    fs.mkdirSync(`${testSrc}/repos/repo30`, err => {
+      if (err) {
+        console.error(err)
+      }
+      // file written successfully
+    })
+    execSync('git --bare init', {
+      stdio: 'pipe', // hide output from git
+      cwd: path.resolve(`${testSrc}/repos/repo30`, '') // where we're cloning the repo to
+    })
+    // Change directory path run git command
+    execSync(`git clone ${testSrc}/repos/repo30`, {
+      stdio: 'pipe', // hide output from git
+      cwd: path.resolve(`${testSrc}/clones`, '') // where we're cloning the repo to
+    })
+    fs.writeFileSync(`${testSrc}/clones/repo30/package.json`, JSON.stringify(repo30Package))
+    fs.writeFileSync(`${testSrc}/clones/repo30/package-lock.json`, JSON.stringify(repo30PackageLocked))
+
+    execSync('npm ci', {
+      stdio: 'pipe', // hide output from git
+      cwd: path.resolve(`${testSrc}/clones/repo13`, '') // where we're cloning the repo to
+    })
+    execSync('npm ci', {
+      stdio: 'pipe', // hide output from git
+      cwd: path.resolve(`${testSrc}/clones/repo13`, '') // where we're cloning the repo to
     })
     execSync('npm ci', {
       stdio: 'pipe', // hide output from git
